honor overridesin do_systemd_enable_load_start()

+++ installer/common.sh
@@ -396,19 +396,23 @@ convert_template() {
 }
 
 do_systemd_enable_load_start() {
-    status=$(systemctl is-enabled $1 2>&1)
-    if [ "$status" == "disabled" ]; then
-        sudo systemctl enable $1 1>/dev/null 2>> $log_file  && logit "${1} systemd unit file enabled" || 
-                    logit "FAILED to enable ${1} systemd unit file" "WARNING"
-    elif [ "$status" == "enabled" ]; then
-        logit "$1 unit file already enabled"
-    elif [[ "$status" =~ "No such file or directory" ]]; then
-        logit "$1 unit file not found" "ERROR" 
+    if [[ ! -f "${override_dir}${1}.service" ]] ; then
+        status=$(systemctl is-enabled $1 2>&1)
+        if [ "$status" == "disabled" ]; then
+            sudo systemctl enable $1 1>/dev/null 2>> $log_file  && logit "${1} systemd unit file enabled" || 
+                        logit "FAILED to enable ${1} systemd unit file" "WARNING"
+        elif [ "$status" == "enabled" ]; then
+            logit "$1 unit file already enabled"
+        elif [[ "$status" =~ "No such file or directory" ]]; then
+            logit "$1 unit file not found" "ERROR" 
+        fi
+        # Will only exectue if systemd script enabled
+        sudo systemctl daemon-reload 2>> $log_file || logit "daemon-reload failed, check logs" "WARNING"
+        sudo systemctl stop $1 >/dev/null 2>&1
+        sudo systemctl start $1 1>/dev/null 2>> $log_file || logit "$1 failed to start, may be normal depending on service/hardware" "WARNING"
+    else
+        logit "Skipping enable and start $1 - override found"
     fi
-    # Will only exectue if systemd script enabled
-    sudo systemctl daemon-reload 2>> $log_file || logit "daemon-reload failed, check logs" "WARNING"
-    sudo systemctl stop $1 >/dev/null 2>&1
-    sudo systemctl start $1 1>/dev/null 2>> $log_file || logit "$1 failed to start, may be normal depending on service/hardware" "WARNING"
 }
 
 process_cmds() {
